pyproject.toml
[project]
name = "inst-android"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.3",
    "aioimaplib>=2.0.1",
    "asyncio>=4.0.0",
    "gmqtt>=0.7.0",
    "pycryptodome>=3.23.0",
    "pycryptodomex>=3.23.0",
    "pydantic>=2.12.5",
    "pyhpke>=0.6.4",
    "pyperclip>=1.11.0",
    "requests>=2.32.5",
    "rnet>=3.0.0rc20",
]

[tool.ruff]
line-length = 80
